fixed calling of semi-unsubscribed clients
When an 'unsubscribe' request is going to be processed, it marks corresponding
'subscription' internal object as PENDING_DELETED. Nobody should start to use
the subscription after that. Current outstanding event deliveries are tracked
by a condition variable to wait for a moment when all of them are finished. Before
this patch there could be a case when new delivery is initiated after the flag
is raised.